+++ apps/common-app/src/apps/reanimated/examples/index.ts
@@ -76,6 +76,10 @@ const BasicNestedLayoutAnimation: React.FC = () =>
   );
 const BokehExample: React.FC = () =>
   React.createElement(require('./BokehExample').default as React.FC);
+const AnimatedTransformExample: React.FC = () =>
+  React.createElement(
+    require('./SharedElementTransitions/AnimatedTransform').default
+  );
 const BorderRadiiExample: React.FC = () =>
   React.createElement(
     require('./SharedElementTransitions/BorderRadii').default
@@ -1336,6 +1340,14 @@ export const EXAMPLES: Record<string, Example> = {
     screen: BorderRadiiExample,
     shouldWork: {
       ios: false, // broken on back gesture
+      android: true,
+    },
+  },
+  AnimatedTransformExample: {
+    title: '[SET] Animated Transform',
+    screen: AnimatedTransformExample,
+    shouldWork: {
+      ios: true,
       android: true,
     },
   },